refactor(cli): move the test harness under test/ and trim module surfaces
- src/test-harness.ts -> test/harness.ts (the one src/-resident test scaffold)
- run-input.ts / projections/simulate.ts: six exports demoted to module-local
- mcp/run-store.ts: cite the real flag (rcd run --select events)
