+++ packages/cli/src/commands/digest.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "vitest";
-import { fixture, runCli, runJson } from "../test-harness";
+import { fixture, runCli, runJson } from "../../test/harness";
 
 describe("digest", () => {
   test("json carries the paragraph and the numbers behind it", async () => {
