+++ packages/cli/src/commands/provenance.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "vitest";
-import { fixture, runCli, runJson } from "../test-harness";
+import { fixture, runCli, runJson } from "../../test/harness";
 
 describe("provenance", () => {
   test("lists the options some layer set, with the winning layer", async () => {
