+++ packages/cli/src/commands/validate.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "vitest";
-import { fixture, runCli, runJson } from "../test-harness";
+import { fixture, runCli, runJson } from "../../test/harness";
 
 describe("validate", () => {
   test("a config Renovate accepts exits 0", async () => {
